[Chore #159361318] Add tests for all basic endpoints
This commit delivers the addition of tests to all basic routes
[delivers #159361318]
