+++ package.json
@@ -7,12 +7,12 @@
   },
   "main": "index.js",
   "scripts": {
-    "test": "babel-node node_modules/mocha/bin/_mocha test/*.js",
+    "test": "babel-node node_modules/mocha/bin/_mocha test/*.js --exit",
     "pretest": "eslint .",
     "lintfix": "eslint . --fix",
     "build": "npm run lintfix && babel src -d dist",
     "migrate": "node ./dist/migration/database.js",
-    "cover": "babel-node ./node_modules/istanbul/lib/cli cover --hook-run-in-context node_modules/mocha/bin/_mocha -- -R spec --timeout 10000 test/*.js",
+    "cover": "babel-node ./node_modules/istanbul/lib/cli cover --hook-run-in-context node_modules/mocha/bin/_mocha -- -R spec --timeout 10000 test/*.js --exit",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "dev": "nodemon ./dist/server",
     "start": "node ./dist/server"
